Add flame tower projectile, hand UI improvements, and ghost hold progress indicator
Flame tower now fires a visible fireball projectile that travels to the
nearest path center point and spawns a fire patch on landing instead of
instantly creating one at the enemy position.

Hand UI tabs moved from a vertical column on the left to a horizontal
row above the cards for easier tapping. Selected cards now lift up 15px
for tactile feedback.

Ghost piece placement shows a radial fill arc during the hold duration,
giving clear visual progress toward placement confirmation.

Also includes misc cleanup: disable goblin camps, FindObjectOfType to
FindFirstObjectByType migration, Unity version bump.

+++ Assets/Scripts/Core/MaterialCache.cs
@@ -71,9 +71,9 @@ namespace TowerDefense.Core
         {
             var mat = new Material(Standard);
             mat.SetFloat("_Mode", 3);
-            mat.SetInt("_SrcBlend", (int)UnityEngine.Rendering.BlendMode.SrcAlpha);
-            mat.SetInt("_DstBlend", (int)UnityEngine.Rendering.BlendMode.OneMinusSrcAlpha);
-            mat.SetInt("_ZWrite", 0);
+            mat.SetFloat("_SrcBlend", (float)UnityEngine.Rendering.BlendMode.SrcAlpha);
+            mat.SetFloat("_DstBlend", (float)UnityEngine.Rendering.BlendMode.OneMinusSrcAlpha);
+            mat.SetFloat("_ZWrite", 0f);
             mat.DisableKeyword("_ALPHATEST_ON");
             mat.EnableKeyword("_ALPHABLEND_ON");
             mat.DisableKeyword("_ALPHAPREMULTIPLY_ON");
